Rename FirestoreSimpleV2 to FirestoreSimple
